♻ :: 기본 키 타입 수정

+++ SpringBoot01/src/main/java/com/example/studyspringboot/domain/feed/domain/repository/FeedRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface FeedRepository extends CrudRepository<Feed, Long> {
+public interface FeedRepository extends CrudRepository<Feed, Integer> {
 
     @Query("select f from Feed f join fetch f.user order by f.createdAt DESC ")
     List<Feed> findAllByJoinFetch();
