+++ SpringBoot01/src/main/java/com/example/studyspringboot/domain/like/presentation/LikeController.java
@@ -18,25 +18,25 @@ public class LikeController {
 
     @ResponseStatus(HttpStatus.CREATED)
     @PostMapping("/{feed-id}")
-    public LikeResponse liked(@PathVariable("feed-id") Long feedId) {
+    public LikeResponse liked(@PathVariable("feed-id") Integer feedId) {
         return likeService.liked(feedId);
     }
 
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/{feed-id}")
-    public LikeResponse deleteLiked(@PathVariable("feed-id") Long feedId) {
+    public LikeResponse deleteLiked(@PathVariable("feed-id") Integer feedId) {
         return likeService.deleteLiked(feedId);
     }
 
     @ResponseStatus(HttpStatus.CREATED)
     @PostMapping("/unlike/{feed-id}")
-    public UnLikeResponse unLiked(@PathVariable("feed-id") Long feedId) {
+    public UnLikeResponse unLiked(@PathVariable("feed-id") Integer feedId) {
         return unLikeService.unLiked(feedId);
     }
 
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/unlike/{feed-id}")
-    public UnLikeResponse deleteUnLiked(@PathVariable("feed-id") Long feedId) {
+    public UnLikeResponse deleteUnLiked(@PathVariable("feed-id") Integer feedId) {
         return unLikeService.deleteUnLiked(feedId);
     }
